+++ packages/testsuite/apps/basic/.wundergraph/wundergraph.server.ts
@@ -1,0 +1,36 @@
+import { configureWunderGraphServer } from '@wundergraph/sdk/server';
+import type { HooksConfig } from './generated/wundergraph.hooks';
+import type { InternalClient } from './generated/wundergraph.internal.client';
+import { buildSchema, GraphQLNonNull, GraphQLObjectType, GraphQLSchema, GraphQLString } from 'graphql/index';
+import { GraphQLExecutionContext } from './generated/wundergraph.server';
+
+export default configureWunderGraphServer<HooksConfig, InternalClient>(() => ({
+	hooks: {
+		queries: {},
+		mutations: {},
+	},
+	graphqlServers: [
+		{
+			apiNamespace: 'embedded',
+			serverName: 'embedded',
+			schema: new GraphQLSchema({
+				query: new GraphQLObjectType<any, GraphQLExecutionContext>({
+					name: 'Query',
+					fields: {
+						clientRequestHeader: {
+							type: GraphQLString,
+							args: {
+								header: {
+									type: new GraphQLNonNull(GraphQLString),
+								},
+							},
+							resolve: async (parent, args, ctx) => {
+								return ctx.wundergraph.clientRequest.headers.get(args.header);
+							},
+						},
+					},
+				}),
+			}),
+		},
+	],
+}));