+++ packages/react-query/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.15](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.14...@wundergraph/react-query@0.8.15) (2023-04-20)
+
+**Note:** Version bump only for package @wundergraph/react-query
+
+## [0.8.14](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.13...@wundergraph/react-query@0.8.14) (2023-04-19)
+
+**Note:** Version bump only for package @wundergraph/react-query
+
+## [0.8.13](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.12...@wundergraph/react-query@0.8.13) (2023-04-19)
+
+**Note:** Version bump only for package @wundergraph/react-query
+
+## [0.8.12](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.11...@wundergraph/react-query@0.8.12) (2023-04-19)
+
+**Note:** Version bump only for package @wundergraph/react-query
+
+## [0.8.11](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.10...@wundergraph/react-query@0.8.11) (2023-04-17)
+
+**Note:** Version bump only for package @wundergraph/react-query
+
 ## [0.8.10](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-query@0.8.9...@wundergraph/react-query@0.8.10) (2023-04-15)
 
 **Note:** Version bump only for package @wundergraph/react-query